Move quick target cards to helper file.
